BACKPORT: ANDROID: dm: Add android verity target
This device-mapper target is virtually a VERITY target. This
target is setup by reading the metadata contents piggybacked
to the actual data blocks in the block device. The signature
of the metadata contents are verified against the key included
in the system keyring. Upon success, the underlying verity
target is setup.

BUG: 27175947

[@acroreiser]: implemented in dm-verity.c; signature verifications
 are omitted because we need this module for compatibility reasons.

Co-Authored-By: Yaroslav Zviezda <acroreiser@gmail.com>